requirements.txt
fastapi
uvicorn
requests
firebase-admin
pyjwt
eth-account
